// src/courses/course-tag.enum.ts
export enum CourseTag {
  Benessere = 'benessere',
  Movimento = 'movimento',
  Energia = 'energia',
  Relax = 'relax',
  Famiglia = 'famiglia',
  Bambini = 'bambini',
  Formazione = 'formazione',
  Lavoro = 'lavoro',
  Creativita = 'creativita',
  Socialita = 'socialita',
  Eventi = 'eventi',
}

export const COURSE_TAG_VALUES = Object.values(CourseTag);